# Optimize Block8x8F low hanging fruit and fix naming