--- a/tests/ImageSharp.Tests/Formats/Jpg/Block8x8FTests.cs
+++ b/tests/ImageSharp.Tests/Formats/Jpg/Block8x8FTests.cs
@@ -5,6 +5,7 @@
 // #define BENCHMARKING
 using System;
 using System.Diagnostics;
+
 using SixLabors.ImageSharp.Formats.Jpeg.Components;
 using SixLabors.ImageSharp.Tests.Formats.Jpg.Utils;
 using SixLabors.ImageSharp.Tests.TestUtilities;
@@ -44,20 +45,20 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             this.Measure(
                 Times,
                 () =>
+                {
+                    var block = default(Block8x8F);
+
+                    for (int i = 0; i < Block8x8F.Size; i++)
                     {
-                        var block = default(Block8x8F);
+                        block[i] = i;
+                    }
 
-                        for (int i = 0; i < Block8x8F.Size; i++)
-                        {
-                            block[i] = i;
-                        }
-
-                        sum = 0;
-                        for (int i = 0; i < Block8x8F.Size; i++)
-                        {
-                            sum += block[i];
-                        }
-                    });
+                    sum = 0;
+                    for (int i = 0; i < Block8x8F.Size; i++)
+                    {
+                        sum += block[i];
+                    }
+                });
             Assert.Equal(sum, 64f * 63f * 0.5f);
         }
 
@@ -69,20 +70,20 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             this.Measure(
                 Times,
                 () =>
+                {
+                    // Block8x8F block = new Block8x8F();
+                    float[] block = new float[64];
+                    for (int i = 0; i < Block8x8F.Size; i++)
                     {
-                        // Block8x8F block = new Block8x8F();
-                        float[] block = new float[64];
-                        for (int i = 0; i < Block8x8F.Size; i++)
-                        {
-                            block[i] = i;
-                        }
+                        block[i] = i;
+                    }
 
-                        sum = 0;
-                        for (int i = 0; i < Block8x8F.Size; i++)
-                        {
-                            sum += block[i];
-                        }
-                    });
+                    sum = 0;
+                    for (int i = 0; i < Block8x8F.Size; i++)
+                    {
+                        sum += block[i];
+                    }
+                });
             Assert.Equal(sum, 64f * 63f * 0.5f);
         }
 
@@ -100,11 +101,11 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             this.Measure(
                 Times,
                 () =>
-                    {
-                        var b = default(Block8x8F);
-                        b.LoadFrom(data);
-                        b.ScaledCopyTo(mirror);
-                    });
+                {
+                    var b = default(Block8x8F);
+                    b.LoadFrom(data);
+                    b.ScaledCopyTo(mirror);
+                });
 
             Assert.Equal(data, mirror);
 
@@ -125,11 +126,11 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             this.Measure(
                 Times,
                 () =>
-                    {
-                        var b = default(Block8x8F);
-                        Block8x8F.LoadFrom(&b, data);
-                        Block8x8F.ScaledCopyTo(&b, mirror);
-                    });
+                {
+                    var b = default(Block8x8F);
+                    Block8x8F.LoadFrom(&b, data);
+                    Block8x8F.ScaledCopyTo(&b, mirror);
+                });
 
             Assert.Equal(data, mirror);
 
@@ -150,11 +151,11 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             this.Measure(
                 Times,
                 () =>
-                    {
-                        var v = default(Block8x8F);
-                        v.LoadFrom(data);
-                        v.ScaledCopyTo(mirror);
-                    });
+                {
+                    var v = default(Block8x8F);
+                    v.LoadFrom(data);
+                    v.ScaledCopyTo(mirror);
+                });
 
             Assert.Equal(data, mirror);
 
@@ -234,7 +235,7 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             this.PrintLinearData(input);
 
             Block8x8F dest = block;
-            dest.NormalizeColorsInplace(255);
+            dest.NormalizeColorsInPlace(255);
 
             float[] array = new float[64];
             dest.ScaledCopyTo(array);
@@ -259,11 +260,11 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             Block8x8F source = CreateRandomFloatBlock(-200, 200, seed);
 
             Block8x8F expected = source;
-            expected.NormalizeColorsInplace(255);
-            expected.RoundInplace();
+            expected.NormalizeColorsInPlace(255);
+            expected.RoundInPlace();
 
             Block8x8F actual = source;
-            actual.NormalizeColorsAndRoundInplaceVector8(255);
+            actual.NormalizeColorsAndRoundInPlaceVector8(255);
 
             this.Output.WriteLine(expected.ToString());
             this.Output.WriteLine(actual.ToString());
@@ -324,12 +325,12 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         [InlineData(1)]
         [InlineData(2)]
         [InlineData(3)]
-        public void RoundInplaceSlow(int seed)
+        public void RoundInPlaceSlow(int seed)
         {
             Block8x8F s = CreateRandomFloatBlock(-500, 500, seed);
 
             Block8x8F d = s;
-            d.RoundInplace();
+            d.RoundInPlace();
 
             this.Output.WriteLine(s.ToString());
             this.Output.WriteLine(d.ToString());
@@ -344,19 +345,26 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         }
 
         [Fact]
-        public void MultiplyInplace_ByOtherBlock()
+        public void MultiplyInPlace_ByOtherBlock()
         {
-            Block8x8F original = CreateRandomFloatBlock(-500, 500, 42);
-            Block8x8F m = CreateRandomFloatBlock(-500, 500, 42);
-
-            Block8x8F actual = original;
-
-            actual.MultiplyInplace(ref m);
-
-            for (int i = 0; i < Block8x8F.Size; i++)
+            static void RunTest()
             {
-                Assert.Equal(original[i] * m[i], actual[i]);
+                Block8x8F original = CreateRandomFloatBlock(-500, 500, 42);
+                Block8x8F m = CreateRandomFloatBlock(-500, 500, 42);
+
+                Block8x8F actual = original;
+
+                actual.MultiplyInPlace(ref m);
+
+                for (int i = 0; i < Block8x8F.Size; i++)
+                {
+                    Assert.Equal(original[i] * m[i], actual[i]);
+                }
             }
+
+            FeatureTestRunner.RunWithHwIntrinsicsFeature(
+                RunTest,
+                HwIntrinsics.AllowAll | HwIntrinsics.DisableAVX);
         }
 
         [Theory]
@@ -396,23 +404,51 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
 
             ReferenceImplementations.DequantizeBlock(&expected, &qt, unzig.Data);
 
-            actual.MultiplyInplace(ref zigQt);
+            actual.MultiplyInPlace(ref zigQt);
 
             this.CompareBlocks(expected, actual, 0);
         }
 
         [Fact]
-        public void MultiplyInplace_ByScalar()
+        public void AddToAllInPlace()
         {
-            Block8x8F original = CreateRandomFloatBlock(-500, 500);
-
-            Block8x8F actual = original;
-            actual.MultiplyInplace(42f);
-
-            for (int i = 0; i < 64; i++)
+            static void RunTest()
             {
-                Assert.Equal(original[i] * 42f, actual[i]);
+                Block8x8F original = CreateRandomFloatBlock(-500, 500);
+
+                Block8x8F actual = original;
+                actual.AddInPlace(42f);
+
+                for (int i = 0; i < 64; i++)
+                {
+                    Assert.Equal(original[i] + 42f, actual[i]);
+                }
             }
+
+            FeatureTestRunner.RunWithHwIntrinsicsFeature(
+                RunTest,
+                HwIntrinsics.AllowAll | HwIntrinsics.DisableAVX);
+        }
+
+        [Fact]
+        public void MultiplyInPlace_ByScalar()
+        {
+            static void RunTest()
+            {
+                Block8x8F original = CreateRandomFloatBlock(-500, 500);
+
+                Block8x8F actual = original;
+                actual.MultiplyInPlace(42f);
+
+                for (int i = 0; i < 64; i++)
+                {
+                    Assert.Equal(original[i] * 42f, actual[i]);
+                }
+            }
+
+            FeatureTestRunner.RunWithHwIntrinsicsFeature(
+                RunTest,
+                HwIntrinsics.AllowAll | HwIntrinsics.DisableAVX);
         }
 
         [Fact]