--- a/tests/ImageSharp.Tests/Formats/Jpg/JpegDecoderTests.Progressive.cs
+++ b/tests/ImageSharp.Tests/Formats/Jpg/JpegDecoderTests.Progressive.cs
@@ -54,8 +54,7 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
             RemoteExecutor.Invoke(
                 RunTest,
                 providerDump,
-                    "Disco")
-                    .Dispose();
+                "Disco").Dispose();
         }
     }
 }